app/requirements.txt
fastapi
uvicorn
langchain
openai
fastapi
python-dotenv